pyproject.toml
[project]
name = "documind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.20",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "llama-index>=0.13.4",
    "llama-index-embeddings-google>=0.4.0",
    "llama-index-llms-google-genai>=0.3.0",
    "llama-index-vector-stores-chroma>=0.5.2",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
